+++ apiserver/config/development/systemRuntime.properties
@@ -16,3 +16,5 @@ alwaysReloadResourcesFile=true
 mergeResourceFiles=false
 addFingerprintToResources=true
 bugsnagApiKey=
+amqp.host=localhost
+amqp.port=5672
