+++ channels-definition/src/main/java/ai/labs/channels/differ/IDifferEndpoint.java
@@ -1,0 +1,5 @@
+package ai.labs.channels.differ;
+
+public interface IDifferEndpoint {
+    void init();
+}
