+++ channels-definition/src/main/java/ai/labs/channels/xmpp/IXmppEndpoint.java
@@ -1,0 +1,6 @@
+package ai.labs.channels.xmpp;
+
+public interface IXmppEndpoint {
+
+    void init();
+}
