+++ channels-impl/src/main/java/ai/labs/channels/differ/model/DifferEventDefinition.java
@@ -1,0 +1,20 @@
+package ai.labs.channels.differ.model;
+
+import ai.labs.models.Context;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotEmpty;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+
+@Getter
+@Setter
+public class DifferEventDefinition {
+    @NotEmpty
+    private List<String> differBotUserId = new LinkedList<>();
+    @NotEmpty
+    private String botIntent;
+    private Map<String, Context> botContext;
+}
