--- xmpp-definition/src/main/java/ai/labs/xmpp/endpoint/IXmppEndpoint.java
@@ -1,8 +1,0 @@
-package ai.labs.xmpp.endpoint;
-
-import org.jivesoftware.smack.chat2.IncomingChatMessageListener;
-
-public interface IXmppEndpoint extends IncomingChatMessageListener {
-
-    void init();
-}
